requirements.txt
eth-brownie
slither-analyzer
scipy
numpy